+++ testdata/cdk/lib/nest/child_stack.go
@@ -1,6 +1,8 @@
-package lib
+package nest
 
 import (
+	"cdk/lib/resource"
+
 	"github.com/aws/aws-cdk-go/awscdk/v2"
 	"github.com/aws/constructs-go/constructs/v10"
 )
@@ -18,10 +20,10 @@ func NewChildStack(scope constructs.Construct, id string, props *ChildStackProps
 
 	stack := awscdk.NewNestedStack(scope, &id, &sprops)
 
-	NewS3Bucket(stack)
-	NewS3DirectoryBucket(stack, props.PjPrefix+"-Child")
-	NewIamGroup(stack)
-	NewCustomResources(stack)
+	resource.NewS3Bucket(stack)
+	resource.NewS3DirectoryBucket(stack, props.PjPrefix+"-Child")
+	resource.NewIamGroup(stack)
+	resource.NewCustomResources(stack)
 
 	NewDescendStack(stack, "Descend", &DescendStackProps{
 		PjPrefix: props.PjPrefix,
