+++ testdata/cdk/lib/nest/descend_stack2.go
@@ -1,16 +1,18 @@
-package lib
+package nest
 
 import (
+	"cdk/lib/resource"
+
 	"github.com/aws/aws-cdk-go/awscdk/v2"
 	"github.com/aws/constructs-go/constructs/v10"
 )
 
-type ChildStack2Props struct {
+type DescendStack2Props struct {
 	awscdk.NestedStackProps
 	PjPrefix string
 }
 
-func NewChildStack2(scope constructs.Construct, id string, props *ChildStack2Props) awscdk.NestedStack {
+func NewDescendStack2(scope constructs.Construct, id string, props *DescendStack2Props) awscdk.NestedStack {
 	var sprops awscdk.NestedStackProps
 	if props != nil {
 		sprops = props.NestedStackProps
@@ -18,12 +20,7 @@ func NewChildStack2(scope constructs.Construct, id string, props *ChildStack2Pro
 
 	stack := awscdk.NewNestedStack(scope, &id, &sprops)
 
-	NewS3Bucket(stack)
-	NewCustomResources(stack)
-
-	NewDescendStack2(stack, "DescendTwo", &DescendStack2Props{
-		PjPrefix: props.PjPrefix,
-	})
+	resource.NewEcr(stack)
 
 	return stack
 }
