+++ testdata/cdk/lib/resource/s3_directory_bucket.go
@@ -1,4 +1,4 @@
-package lib
+package resource
 
 import (
 	"strings"
